Set up email (partially)
I added a feature where upon submitting an order, the order is submitted to the owner of the site via email. In addition, the customer is sent an automatic reply.

I used emailjs for this.

Note to self:
Right now, the order items and price are not being submitted in the email. I think it is because the inputs are being assigned values via JS and not by the user.